requirements.txt
pytest==7.4.0
selenium==4.10.0
allure-pytest==2.13.2
requests==2.32.4